+++ test/lib/mayaUsd/render/mayaToHydra/cpp/testMergingSceneIndex.cpp
@@ -33,7 +33,7 @@ TEST(FlowViewport, mergingSceneIndex)
 {
     // The Flow Viewport custom merging scene index is in the scene index tree.
     const auto& sceneIndices = GetTerminalSceneIndices();
-    ASSERT_GT(sceneIndices.size(), static_cast<size_t>(0));
+    ASSERT_GT(sceneIndices.size(), 0u);
 
     auto isFvpMergingSceneIndex = SceneIndexDisplayNamePred(
         "Flow Viewport Merging Scene Index");
